pyproject.toml
[project]
name = "shopify_psrestful"
version = "0.1.0"
description = "PSRESTful scripts for Shopify, helping promotional products distributors to integrate with Shopify using PromoStandards"
authors = [
    { name = "Edilio Gallardo", email = "edilio@gallardo.solutions" }
]
packages = [
    { include = "shopify_psrestful", from = "src" },
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

[tool.poetry.dependencies]
tenacity = "^8.2.3"
shopifyapi = { version = "==12.7.0"}
python-dotenv = { version = "^1.0.0"}
psdomain = { git = "git@github.com:GallardoSolutions/psdomain.git", rev = "v0.0.67" }

[tool.poetry]
packages = [{ include = "shopify_psrestful", from = "src" }]

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
